requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2020.10.14=0
certifi=2020.6.20=pyhd3eb1b0_3
cycler=0.10.0=pypi_0
dbus=1.13.18=hb2f20db_0
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
matplotlib=3.3.2=pypi_0
ncurses=6.2=he6710b0_1
numpy=1.19.4=pypi_0
openssl=1.1.1h=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=8.0.1=pypi_0
pip=20.2.4=py38h06a4308_0
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
setuptools=50.3.0=py38h06a4308_1
sip=4.19.13=py38he6710b0_0
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
